requirements.txt
# pip install -r requirements.txt

requests==2.28.2
pytest==7.2.2
pytest-html==3.2.0
pytest-ordering==0.6
PyYAML==6.0
